Update user prompt add post hook
Post rewrites `setup.py` to more closely match the creation of the file
by paster.

Once the variables in `cookiecutter.json` are set, they can't
be changed. As a result, the processing Paster was doing to handle the
creation of the `project_shortname` and `plugin_class_name` have been
moved to the `cookiecutter.json` which means the user will be prompted
for values but the default values will be correct if they are left
alone.

The Paster script did some work to ensure that the user provided
keywords had at one, and only one, instance of 'CKAN.' This is handled
in the `post_gen` hook. Inorder to do this the the provided keywords
are parsed and changed as needed then the `setup.py` is recreated with
the new values. This process also double checks that `project_shortname`
and `plugin_class_name` are correct and make adjstments as needed.